requirements.txt
# FTW-VLM dependencies
# Install with: pip install -r requirements.txt

# Core geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
geojson>=3.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Gemini API
google-genai>=0.3.0

# Optional — install for best skeleton thinning quality (Zhang-Suen)
# pip install opencv-contrib-python
